Check feed active state
