+++ nextjs/components/NavBar/Desktop/NavLabel/index.module.scss
@@ -11,6 +11,5 @@
 
   padding: 0.5rem;
   width: 100%;
-  margin: 1rem 0 0;
   transition: 0.1s ease-in color, 0.1s ease-in background;
 }
